More updates to code structure/placement
