+++ routes/index.js
@@ -1,58 +1,32 @@
 const express = require('express');
 const passport = require('passport');
-const Account = require('../models/account');
 const router = express.Router();
+const account_controller = require('../controllers/accountController');
 
-// GET home/goals page
-router.get('/', function(req, res) {
-    res.redirect('/goals');
-});
+router.get('/', account_controller.account_home_get);
+
+/** Registration Routes **/
 
 // GET request for registration page
-router.get('/register', function (req, res) {
-    if(req.user){
-        res.redirect('/');
-    } else {
-        res.render('register');
-    }
-});
+router.get('/register', account_controller.account_registration_get);
 
 // POST request for creating an account
-router.post('/register', function (req, res) {
-    Account.register(new Account({username: req.body.username}),
-        req.body.password, function (err, account) {
+router.post('/register', account_controller.account_registration_post);
 
-        if (err || (req.body.password !== req.body.re_password)) {
-            // ToDo: Currently just refreshes page if invalid. Add response indicating error to user
-            return res.render('register');
-        }
 
-        passport.authenticate('local')(req, res, function () {
-            res.redirect('/goals', {user: req.user});
-        });
-    });
-});
+/** Login Routes **/
 
 // GET request for login page
-router.get('/login', function (req, res) {
-    if (req.user){
-        res.redirect('/');
-    } else {
-        res.render('login', {user: req.user});
-    }
-
-});
+router.get('/login', account_controller.account_login_get);
 
 // POST request for logging in
 router.post('/login', passport.authenticate('local',
-    {successRedirect: '/', failureRedirect: '/login', failureFlash: true}
-));
+    {successRedirect: '/goals', failureRedirect: '/login', failureFlash: true}));
 
+
+/** Logout Routes **/
 
 // GET request for logging out
-router.get('/logout', function (req, res) {
-    req.logout();
-    res.redirect('/login');
-});
+router.post('/logout', account_controller.account_logout_post);
 
 module.exports = router;
